# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Soundwave Music Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface User {
  id: number;
  clerkId: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UserUpdate {
  /** @minLength 1 */
  username?: string;
  displayName?: string;
}

export interface Artist {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isVerified?: boolean;
  albumCount: number;
  createdAt: string;
}

export interface Album {
  id: number;
  title: string;
  artistId: number;
  artistName: string;
  /** @nullable */
  coverUrl?: string | null;
  uploaderUsername: string;
  uploaderClerkId?: string;
  songCount: number;
  createdAt: string;
}

export interface ArtistDetail {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isVerified?: boolean;
  albumCount: number;
  albums: Album[];
  createdAt: string;
}

export interface Song {
  id: number;
  title: string;
  albumId: number;
  albumTitle: string;
  artistId: number;
  artistName: string;
  /** @nullable */
  coverUrl?: string | null;
  uploaderUsername: string;
  uploaderClerkId?: string;
  /**
     * Duration in seconds
     * @nullable
     */
  duration: number | null;
  objectPath: string;
  /** @nullable */
  trackNumber?: number | null;
  createdAt: string;
}

export interface AlbumDetail {
  id: number;
  title: string;
  artistId: number;
  artistName: string;
  /** @nullable */
  coverUrl?: string | null;
  uploaderUsername: string;
  uploaderClerkId?: string;
  songs: Song[];
  createdAt: string;
}

export interface SongInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  objectPath: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  trackNumber?: number | null;
}

export interface AlbumInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  artistName: string;
  /** @nullable */
  coverObjectPath?: string | null;
  /** @minItems 1 */
  songs: SongInput[];
}

export interface Playlist {
  id: number;
  name: string;
  isPrivate: boolean;
  songCount: number;
  ownerUsername: string;
  /** @nullable */
  coverUrl?: string | null;
  createdAt: string;
}

export interface PlaylistDetail {
  id: number;
  name: string;
  isPrivate: boolean;
  ownerUsername: string;
  /** @nullable */
  coverUrl?: string | null;
  songs: Song[];
  createdAt: string;
}

export interface PlaylistInput {
  /** @minLength 1 */
  name: string;
  isPrivate?: boolean;
}

export interface PlaylistUpdate {
  /** @minLength 1 */
  name?: string;
  isPrivate?: boolean;
}

export interface PlaylistSongInput {
  songId: number;
}

export interface FeaturedData {
  recentAlbums: Album[];
  popularArtists: Artist[];
}

export interface PlatformStats {
  totalAlbums: number;
  totalSongs: number;
  totalArtists: number;
  totalUsers: number;
}

export type ListArtistsParams = {
/**
 * Filter by artist name
 */
search?: string;
limit?: number;
offset?: number;
};

export type ListAlbumsParams = {
/**
 * Filter by album or artist name
 */
search?: string;
/**
 * Filter by artist
 */
artistId?: number;
limit?: number;
offset?: number;
};

export type ListSongsParams = {
/**
 * Filter by song or artist name
 */
search?: string;
albumId?: number;
limit?: number;
offset?: number;
};