# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Soundwave Music Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /albums:
    get:
      tags: [Albums]
      operationId: listAlbums
      summary: List all albums
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Filter by album or artist name
        - name: artistId
          in: query
          schema:
            type: integer
          description: Filter by artist
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"

    post:
      tags: [Albums]
      operationId: createAlbum
      summary: Upload a new album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumInput"
      responses:
        "201":
          description: Album created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /albums/{id}:
    get:
      tags: [Albums]
      operationId: getAlbum
      summary: Get album details with songs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Album with songs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumDetail"
        "404":
          description: Album not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      tags: [Albums]
      operationId: deleteAlbum
      summary: Delete an album (uploader only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /albums/{id}/songs:
    post:
      tags: [Albums]
      operationId: addSongToAlbum
      summary: Add a song to an album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "201":
          description: Song added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Album not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Album:
      type: object
      required: [id, title, artistId, artistName, uploaderUsername, songCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artistId:
          type: integer
        artistName:
          type: string
        coverUrl:
          type: ["string", "null"]
        uploaderUsername:
          type: string
        uploaderClerkId:
          type: string
        songCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    AlbumDetail:
      type: object
      required: [id, title, artistId, artistName, uploaderUsername, songs, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artistId:
          type: integer
        artistName:
          type: string
        coverUrl:
          type: ["string", "null"]
        uploaderUsername:
          type: string
        uploaderClerkId:
          type: string
        songs:
          type: array
          items:
            $ref: "#/components/schemas/Song"
        createdAt:
          type: string
          format: date-time
    AlbumInput:
      type: object
      required: [title, artistName, songs]
      properties:
        title:
          type: string
          minLength: 1
        artistName:
          type: string
          minLength: 1
        coverObjectPath:
          type: ["string", "null"]
        songs:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SongInput"
    Song:
      type: object
      required: [id, title, albumId, albumTitle, artistId, artistName, uploaderUsername, duration, objectPath, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        albumId:
          type: integer
        albumTitle:
          type: string
        artistId:
          type: integer
        artistName:
          type: string
        coverUrl:
          type: ["string", "null"]
        uploaderUsername:
          type: string
        uploaderClerkId:
          type: string
        duration:
          type: ["integer", "null"]
          description: Duration in seconds
        objectPath:
          type: string
        trackNumber:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    SongInput:
      type: object
      required: [title, objectPath]
      properties:
        title:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        duration:
          type: ["integer", "null"]
        trackNumber:
          type: ["integer", "null"]
...