# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Soundwave Music Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: number;
  title: string;
  artistId: number;
  artistName: string;
  /** @nullable */
  coverUrl?: string | null;
  uploaderUsername: string;
  uploaderClerkId?: string;
  songCount: number;
  createdAt: Date;
}