# lib/api-zod/src/generated/types/songInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Soundwave Music Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SongInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  objectPath: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  trackNumber?: number | null;
}